+++ example/lib/screens/screens.dart
@@ -14,6 +14,7 @@ export 'widgets/bottom_sheet_screen.dart';
 export 'widgets/expansion_list_item_screen.dart';
 export 'widgets/filter_pill_screen.dart';
 export 'widgets/input_screen.dart';
+export 'widgets/slider_screen.dart';
 
 // tools
 export 'theme_editor/theme_editor_screen.dart';