+++ example/lib/screens/widgets/button_screen.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/widgets.dart';
+
 import 'package:prime_flutter/prime_flutter.dart';
 
 class ButtonScreen extends StatelessWidget {