+++ lib/prime_flutter.dart
@@ -23,4 +23,5 @@ export 'src/widgets/prime_icons.dart';
 export 'src/widgets/prime_page_route.dart';
 export 'src/widgets/prime_scaffold.dart';
 export 'src/widgets/progress.dart';
+export 'src/widgets/slider.dart';
 export 'src/widgets/tab_bar.dart';